requirements.txt
telepot
pyyaml
